chore: move ExecutionsTable
