+++ src/components/organisms/EntityDetails/EntityDetailsContent/EntityDetailsContentTabs.tsx
@@ -6,7 +6,7 @@ import useTrackTimeAnalytics from '@hooks/useTrackTimeAnalytics';
 
 import {CLICommands, MetricsBarChart} from '@molecules';
 
-import ExecutionsTable from '@organisms/EntityDetails/EntityDetailsContent/ExecutionsTable';
+import ExecutionsTable from '@organisms/EntityDetails/ExecutionsTable';
 
 import {useEntityDetailsPick} from '@store/entityDetails';
